Made most small requested changes, added bts7xxx_common_impl for private common functions
