+++ libraries/ms-drivers/src/bts7xxx_common.c
@@ -1,12 +1,11 @@
-// Common helper functions for BTS7XXX-series load switches
-
 #include "bts7xxx_common.h"
 
-// Fault restart delay is the same across both the BTS7040 and the BTS7200
-#define BTS7XXX_FAULT_RESTART_DELAY_MS 110
-#define BTS7XXX_FAULT_RESTART_DELAY_US (BTS7XXX_FAULT_RESTART_DELAY_MS * 1000)
-
 StatusCode bts7xxx_enable_pin(Bts7xxxEnablePin *pin) {
+  // Don't allow enabling pin on fault
+  if (pin->fault_in_progress) {
+    return STATUS_CODE_INTERNAL_ERROR;
+  }
+
   if (pin->pin_type == BTS7XXX_PIN_STM32) {
     return gpio_set_state(pin->enable_pin_stm32, GPIO_STATE_HIGH);
   } else {
@@ -14,7 +13,6 @@ StatusCode bts7xxx_enable_pin(Bts7xxxEnablePin *pin) {
   }
 }
 
-// Broad function to disable the pin passed in.
 StatusCode bts7xxx_disable_pin(Bts7xxxEnablePin *pin) {
   if (pin->pin_type == BTS7XXX_PIN_STM32) {
     return gpio_set_state(pin->enable_pin_stm32, GPIO_STATE_LOW);
@@ -23,7 +21,6 @@ StatusCode bts7xxx_disable_pin(Bts7xxxEnablePin *pin) {
   }
 }
 
-// Broad function to get whether the pin passed in is enabled.
 StatusCode bts7xxx_get_pin_enabled(Bts7xxxEnablePin *pin) {
   if (pin->pin_type == BTS7XXX_PIN_STM32) {
     GpioState pin_state;
@@ -34,34 +31,4 @@ StatusCode bts7xxx_get_pin_enabled(Bts7xxxEnablePin *pin) {
     pca9539r_gpio_get_state(pin->enable_pin_pca9539r, &pin_state);
     return (pin_state == PCA9539R_GPIO_STATE_HIGH);
   }
-}
-
-// Broad pin soft timer cb without re-enabling the pin
-void bts7xxx_fault_handler_cb(SoftTimerId timer_id, void *context) {
-  Bts7xxxEnablePin *pin = context;
-  pin->fault_in_progress = false;
-  pin->fault_timer_id = SOFT_TIMER_INVALID_TIMER;
-}
-
-// Broad pin re-enable soft timer cb
-void bts7xxx_fault_handler_enable_cb(SoftTimerId timer_id, void *context) {
-  Bts7xxxEnablePin *pin = context;
-  bts7xxx_fault_handler_cb(timer_id, context);
-  bts7xxx_enable_pin(pin);
-}
-
-// Helper function to clear fault on a given pin
-StatusCode bts7xxx_handle_fault_pin(Bts7xxxEnablePin *pin) {
-  if (!pin->fault_in_progress) {
-    pin->fault_in_progress = true;
-    if (bts7xxx_get_pin_enabled(pin)) {
-      status_ok_or_return(bts7xxx_disable_pin(pin));
-      soft_timer_start(BTS7XXX_FAULT_RESTART_DELAY_US, bts7xxx_fault_handler_enable_cb, pin,
-                       &pin->fault_timer_id);
-    } else {
-      soft_timer_start(BTS7XXX_FAULT_RESTART_DELAY_US, bts7xxx_fault_handler_cb, pin,
-                       &pin->fault_timer_id);
-    }
-  }
-  return STATUS_CODE_OK;
 }
