+++ libraries/ms-drivers/test/test_bts_7200_load_switch.c
@@ -1,4 +1,5 @@
 #include "bts_7200_load_switch.h"
+#include "controller_board_pins.h"
 #include "delay.h"
 #include "interrupt.h"
 #include "log.h"
@@ -6,11 +7,6 @@
 #include "unity.h"
 
 #define TEST_I2C_PORT I2C_PORT_2
-
-#define TEST_CONFIG_PIN_I2C_SCL \
-  { GPIO_PORT_B, 10 }
-#define TEST_CONFIG_PIN_I2C_SDA \
-  { GPIO_PORT_B, 11 }
 
 // 1.6k resistor
 #define BTS7200_TEST_RESISTOR 1600
@@ -75,9 +71,9 @@ void setup_test(void) {
   adc_init(ADC_MODE_SINGLE);
 
   I2CSettings i2c_settings = {
-    .speed = I2C_SPEED_FAST,         //
-    .sda = TEST_CONFIG_PIN_I2C_SDA,  //
-    .scl = TEST_CONFIG_PIN_I2C_SCL,  //
+    .speed = I2C_SPEED_FAST,                //
+    .sda = CONTROLLER_BOARD_ADDR_I2C2_SDA,  //
+    .scl = CONTROLLER_BOARD_ADDR_I2C2_SCL,  //
   };
   i2c_init(TEST_I2C_PORT, &i2c_settings);
 
@@ -384,7 +380,7 @@ void test_bts_7200_current_sense_get_measurement_stm32_valid(void) {
 
   uint16_t reading0 = 0, reading1 = 0;
   TEST_ASSERT_OK(bts_7200_get_measurement(&s_storage, &reading0, &reading1));
-  LOG_DEBUG("STM32 readings: %d, %d\r\n", reading0, reading1);
+  LOG_DEBUG("STM32 readings: %d, %d\n", reading0, reading1);
 }
 
 // Same, but with pca9539r initialization.
@@ -412,7 +408,7 @@ void test_bts_7200_current_sense_get_measurement_pca9539r_valid(void) {
 
   uint16_t reading0 = 0, reading1 = 0;
   TEST_ASSERT_OK(bts_7200_get_measurement(&s_storage, &reading0, &reading1));
-  LOG_DEBUG("PCA9539R readings: %d, %d\r\n", reading0, reading1);
+  LOG_DEBUG("PCA9539R readings: %d, %d\n", reading0, reading1);
 }
 
 // Test that bts_7200_stop returns true only when it stops a timer
